LPS-116679 Move CSV writers out of DataFactory into a holder which is made available in the Freemarker context
